# api-docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "School Bus Tracking API - v2 Dynamic",
        "description": "Refactored API supporting dynamic Trip orchestration between Drivers, Buses, and Routes.",
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:5000/api"
        }
    ],
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
...
            "UserCreate": {
                "type": "object",
                "required": [
                    "name",
                    "email",
                    "password",
                    "role"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "DRIVER",
                            "PARENT"
                        ]
                    }
                }
            },
...
            "Attendance": {
                "type": "object",
                "required": [
                    "studentId",
                    "tripId",
                    "status"
                ],
                "properties": {
                    "studentId": {
                        "type": "string"
                    },
                    "tripId": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "PICKED_UP",
                            "DROPPED_OFF"
                        ]
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            }
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ],
    "paths": {
...
        "/api/admin/users": {
            "post": {
                "summary": "Create User (Admin)",
                "tags": [
                    "Admin"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
        "/api/admin/trip-history": {
            "get": {
                "summary": "Get Journey History (Admin)",
                "tags": [
                    "Admin Tracking"
                ],
                "parameters": [
                    {
                        "name": "driverId",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "busId",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "routeId",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/admin/trip-location/{tripId}": {
            "get": {
                "summary": "Get Bus Location (Admin)",
                "tags": [
                    "Admin Tracking"
                ],
                "parameters": [
                    {
                        "name": "tripId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/admin/buses/{id}/status": {
            "put": {
                "summary": "Update Bus Status (Admin)",
                "tags": [
                    "Admin"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "isActive": {
                                        "type": "boolean"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/admin/students/{id}/route": {
            "put": {
                "summary": "Reassign Student Route (Admin)",
                "tags": [
                    "Admin"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "routeId": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/admin/students/{id}/bus": {
            "put": {
                "summary": "Reassign Student Bus (Admin)",
                "tags": [
                    "Admin"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "busId": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
        "/api/attendance": {
            "post": {
                "summary": "Mark Attendance",
                "tags": [
                    "Attendance"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Attendance"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        }
    }
}